Split files for external and utility funcs
