+++ src/animate.js
@@ -1,0 +1,8 @@
+let animate;
+if (typeof Velocity === 'function') {
+  animate = Velocity;
+} else {
+  animate = $.Velocity;
+}
+
+export default animate;
